refactoring: return erros and feat: add email validator
