+++ package.json
@@ -16,6 +16,7 @@
         "bcryptjs": "^2.4.3",
         "dd-trace": "^2.12.1",
         "dotenv": "^16.0.1",
+        "email-validator": "^2.0.4",
         "express": "^4.17.1",
         "express-async-errors": "^3.1.1",
         "jsonwebtoken": "^8.5.1",
