+++ src/app.ts
@@ -1,27 +1,15 @@
 import "dotenv/config";
 import "./database/mongoDB";
 import "express-async-errors";
-import express, { NextFunction, Request, Response } from "express";
+import "./infra/tracer";
+import express from "express";
 import { routes } from "./routes";
-import { AppError } from "./errors/AppError";
+import { ReturnErrorsCustomized } from "./utils/ReturnErrorsCustomized";
 
 const app = express();
 app.use(express.json());
 app.use(routes);
 
-app.use(
-    (err: Error, request: Request, response: Response, next: NextFunction) => {
-        if (err instanceof AppError) {
-            return response
-                .status(err.statusCode)
-                .json({ message: err.message });
-        }
-
-        return response.status(500).json({
-            status: "error",
-            message: `Internal server error - ${err.message}`,
-        });
-    }
-);
+app.use(new ReturnErrorsCustomized().execute);
 
 export { app };
