+++ src/routes/apis/githubProfile.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { GithubProfileController } from "../useCase/githubProfile/GithubProfileController";
+import { GithubProfileController } from "../../useCase/githubProfile/GithubProfileController";
 
 const githubProfileRoute = Router();
