+++ src/routes/index.ts
@@ -1,8 +1,8 @@
 import { Router } from "express";
-import { authUserRoute } from "./authUser.routes";
-import { createUserRoute } from "./createUser.routes";
-import { githubProfileRoute } from "./githubProfile.routes";
 import { Auth } from "./middleware/Auth";
+import { githubProfileRoute } from "./apis/githubProfile.routes";
+import { authUserRoute } from "./user/authUser.routes";
+import { createUserRoute } from "./user/createUser.routes";
 
 const routes = Router();
 const authMiddleware = new Auth();
