+++ src/routes/user/authUser.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { AuthUserController } from "../useCase/Auth/AuthUserController";
+import { AuthUserController } from "../../useCase/Auth/AuthUserController";
 
 const authUserRoute = Router();
