+++ src/routes/user/createUser.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { CreateUserController } from "../useCase/createUser/CreateUserController";
+import { CreateUserController } from "../../useCase/createUser/CreateUserController";
 
 const createUserRoute = Router();
