+++ src/server.ts
@@ -1,4 +1,3 @@
-import "./infra/tracer";
 import { app } from "./app";
 import { logger } from "./errors/Winston";
