+++ src/utils/PasswordHash.ts
@@ -1,5 +1,5 @@
 import { hash } from "bcryptjs";
-export class passwordHash {
+export class PasswordHash {
     async hash(password: string) {
         return hash(password, 8);
     }
